// src/services/openapi/notification-service/api.ts
export interface GenericBucketRequestDtoReminderTemplateFilterRequestDto {
    /**
     * 
     * @type {PageRequestDto}
     * @memberof GenericBucketRequestDtoReminderTemplateFilterRequestDto
     */
    page?: PageRequestDto;
    /**
     * 
     * @type {SortRequestDto}
     * @memberof GenericBucketRequestDtoReminderTemplateFilterRequestDto
     */
    sortList?: SortRequestDto;
    /**
     * 
     * @type {SearchDetailRequestDto}
     * @memberof GenericBucketRequestDtoReminderTemplateFilterRequestDto
     */
    searchDetail?: SearchDetailRequestDto;
    /**
     * 
     * @type {ReminderTemplateFilterRequestDto}
     * @memberof GenericBucketRequestDtoReminderTemplateFilterRequestDto
     */
    filter?: ReminderTemplateFilterRequestDto;
}
// ...
export interface GenericBucketRequestDtoTransactionReminderTemplateSubmissionFilterRequestDto {
    /**
     * 
     * @type {PageRequestDto}
     * @memberof GenericBucketRequestDtoTransactionReminderTemplateSubmissionFilterRequestDto
     */
    page?: PageRequestDto;
    /**
     * 
     * @type {SortRequestDto}
     * @memberof GenericBucketRequestDtoTransactionReminderTemplateSubmissionFilterRequestDto
     */
    sortList?: SortRequestDto;
    /**
     * 
     * @type {SearchDetailRequestDto}
     * @memberof GenericBucketRequestDtoTransactionReminderTemplateSubmissionFilterRequestDto
     */
    searchDetail?: SearchDetailRequestDto;
    /**
     * 
     * @type {TransactionReminderTemplateSubmissionFilterRequestDto}
     * @memberof GenericBucketRequestDtoTransactionReminderTemplateSubmissionFilterRequestDto
     */
    filter?: TransactionReminderTemplateSubmissionFilterRequestDto;
}
// ...
export interface PageRequestDto {
    /**
     * 
     * @type {number}
     * @memberof PageRequestDto
     */
    noPage?: number;
    /**
     * 
     * @type {number}
     * @memberof PageRequestDto
     */
    itemPerPage?: number;
}
// ...
export interface ReminderReceiverRequestDto {
    /**
     * 
     * @type {number}
     * @memberof ReminderReceiverRequestDto
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof ReminderReceiverRequestDto
     */
    divisionCode: string;
    /**
     * 
     * @type {string}
     * @memberof ReminderReceiverRequestDto
     */
    roleCode: string;
    /**
     * 
     * @type {string}
     * @memberof ReminderReceiverRequestDto
     */
    positionCode: string;
    /**
     * 
     * @type {boolean}
     * @memberof ReminderReceiverRequestDto
     */
    isActive?: boolean;
}
// ...
export interface ReminderTemplateFilterRequestDto {
    /**
     * 
     * @type {string}
     * @memberof ReminderTemplateFilterRequestDto
     */
    templateCode?: string;
    /**
     * 
     * @type {Array<string>}
     * @memberof ReminderTemplateFilterRequestDto
     */
    templateType?: Array<string>;
    /**
     * 
     * @type {boolean}
     * @memberof ReminderTemplateFilterRequestDto
     */
    isActive?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof ReminderTemplateFilterRequestDto
     */
    isSendEmail?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof ReminderTemplateFilterRequestDto
     */
    isSendLos?: boolean;
    /**
     * 
     * @type {string}
     * @memberof ReminderTemplateFilterRequestDto
     */
    startDate?: string;
    /**
     * 
     * @type {string}
     * @memberof ReminderTemplateFilterRequestDto
     */
    endDate?: string;
}
// ...
export interface RequestByIdDtoString {
    /**
     * 
     * @type {string}
     * @memberof RequestByIdDtoString
     */
    id?: string;
}
// ...
export interface SaveBeforeEditRequestDto {
    /**
     * 
     * @type {number}
     * @memberof SaveBeforeEditRequestDto
     */
    reminderTemplateId?: number;
}
/**
 * 
 * @export
 * @interface SaveBeforeEditResponseDto
 */
export interface SaveBeforeEditResponseDto {
    /**
     * 
     * @type {string}
     * @memberof SaveBeforeEditResponseDto
     */
    bucketProcessId?: string;
}
/**
 * 
 * @export
 * @interface SearchDetailRequestDto
 */
export interface SearchDetailRequestDto {
    /**
     * 
     * @type {string}
     * @memberof SearchDetailRequestDto
     */
    key?: string;
    /**
     * 
     * @type {string}
     * @memberof SearchDetailRequestDto
     */
    value?: string;
}
/**
 * 
 * @export
 * @interface SortRequestDto
 */
export interface SortRequestDto {
    /**
     * 
     * @type {string}
     * @memberof SortRequestDto
     */
    columnName?: string;
    /**
     * 
     * @type {string}
     * @memberof SortRequestDto
     */
    sortType?: string;
}
// ...
export interface TransactionReminderTemplateSubmissionFilterRequestDto {
    /**
     * 
     * @type {string}
     * @memberof TransactionReminderTemplateSubmissionFilterRequestDto
     */
    bucketProcessId?: string;
    /**
     * 
     * @type {Array<object>}
     * @memberof TransactionReminderTemplateSubmissionFilterRequestDto
     */
    templateTypes?: Array<object>;
    /**
     * 
     * @type {Array<object>}
     * @memberof TransactionReminderTemplateSubmissionFilterRequestDto
     */
    statuses?: Array<object>;
    /**
     * 
     * @type {string}
     * @memberof TransactionReminderTemplateSubmissionFilterRequestDto
     */
    templateCode?: string;
}
// ...
export class ReminderTemplateControllerApi extends BaseAPI {
    /**
     * 
     * @summary API to get detail notification template
     * @param {RequestByIdDtoString} requestByIdDtoString 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof ReminderTemplateControllerApi
     */
    public getDetilReminderTemplate(requestByIdDtoString: RequestByIdDtoString, options?: any) {
        return ReminderTemplateControllerApiFp(this.configuration).getDetilReminderTemplate(requestByIdDtoString, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary API to get list notification template
     * @param {GenericBucketRequestDtoReminderTemplateFilterRequestDto} genericBucketRequestDtoReminderTemplateFilterRequestDto 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof ReminderTemplateControllerApi
     */
    public getReminderTemplate(genericBucketRequestDtoReminderTemplateFilterRequestDto: GenericBucketRequestDtoReminderTemplateFilterRequestDto, options?: any) {
        return ReminderTemplateControllerApiFp(this.configuration).getReminderTemplate(genericBucketRequestDtoReminderTemplateFilterRequestDto, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary API to get submission list for approval
     * @param {GenericBucketRequestDtoTransactionReminderTemplateSubmissionFilterRequestDto} genericBucketRequestDtoTransactionReminderTemplateSubmissionFilterRequestDto 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof ReminderTemplateControllerApi
     */
    public getReminderTemplateSubmission(genericBucketRequestDtoTransactionReminderTemplateSubmissionFilterRequestDto: GenericBucketRequestDtoTransactionReminderTemplateSubmissionFilterRequestDto, options?: any) {
        return ReminderTemplateControllerApiFp(this.configuration).getReminderTemplateSubmission(genericBucketRequestDtoTransactionReminderTemplateSubmissionFilterRequestDto, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary API to get submission detail for approval
     * @param {RequestByIdDtoString} requestByIdDtoString 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof ReminderTemplateControllerApi
     */
    public getReminderTemplateSubmissionDetail(requestByIdDtoString: RequestByIdDtoString, options?: any) {
        return ReminderTemplateControllerApiFp(this.configuration).getReminderTemplateSubmissionDetail(requestByIdDtoString, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary API to create transaction from existing reminder template
     * @param {SaveBeforeEditRequestDto} saveBeforeEditRequestDto 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof ReminderTemplateControllerApi
     */
    public saveBeforeEditReminderTemplate(saveBeforeEditRequestDto: SaveBeforeEditRequestDto, options?: any) {
        return ReminderTemplateControllerApiFp(this.configuration).saveBeforeEditReminderTemplate(saveBeforeEditRequestDto, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary API to save transaction reminder template
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof ReminderTemplateControllerApi
     */
    public saveTransactionReminderTemplate(options?: any) {
        return ReminderTemplateControllerApiFp(this.configuration).saveTransactionReminderTemplate(options).then((request) => request(this.axios, this.basePath));
    }

}